Merged PR 2056066: Surface underlying Java exception on failure
There are multiple places in the Java extension where an exception might occur in JVM, but we don't check it. Instead, we throw a c++ runtime_error exception with a fixed error message, that doesn't surface the underline issue.

This change replaces those instances with a call to JniHelper::ThrowOnJavaException, which checks if a Java exception occured

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/2056066

Additional changes:
1. Cleaned up dead code JniHelper::HandleJavaException in JniHelper.cpp
2. Updated JniHelper::ThrowOnJavaException to check if a Java exception was thrown using ExceptionCheck() instead of ExceptionOccurred().
In addition to ExceptionCheck being more semantically accurate to what we want to do, it also doesn't take a local reference on the exception object (unlike ExceptionOccurred)
3. In JniHelper::LogJavaException, replaced manually calling Java exception class' method printStackTrace, use the jni function ExceptionDescribe()