Merged PR 2179347: Add JAVA_OPTS and remove aliases
Add JAVA_OPTS
Added a string split function to help with splitting java_opts

JAVA_OPTS is an environment variable that can be used to pass options to Java. This can be used for passing in options for things such as heap size, debugging, etc. JAVA_OPTS can be passed into the language when creating the language in the ENVIRONMENT_VARIABLES string as part of the dictionary of vars.

We do NOT accept the options "java.class.path" and "file.encoding" because we explicitly set these in the extension to our required values. If these are passed into JAVA_OPTS the java initialization will throw an error.