Merged PR 2502316: Restore and Build googletest framework using CMake
**[Background]**
With the move to use Rcpp for the RExtension, we needed to use Mingw compiler on Windows for its compilation. This is required even for the RExtension-test project because the test executable needs to include Rcpp for advanced tests that query the R environment to verify RExtension's work such as parameter initialization etc. That change to move RExtension-test will be in a subsequent PR.

**[This change]**
Moving RExtension-test to MingW also means the gtest libraries (that it depends on) need to be compiled with MingW. This PR is to do that.
1) Get the latest source from googletest-src master using CMake ExternalProject include => CMakeLists.txt.in
2) Build the googletest project in this repository and make it available in the .build\googletest\\${PLATFORM}\lib directory for the test project to use.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/2502316