Merged PR 2753281: [RExtension] Implement Execute API, its tests and few nits
This PR implements the **Execute** external language API for the RExtension.
The implementation involves 3 major steps :
1) Make the InputDataSet available in the embedded R environment,
2) Evaluate the R script and
3) Retrieve OutputDataSet from the embedded R environment if any.
It adds base class **RDataSet** and two of its derived classes :  **RInputDataSet** and **ROutputDataSet** to achieve this.
Renames Column to **RColumn** and refactors RSession to use these new classes.

The supported column types for an InputDataSet are integer, numeric, logical and character. **RExtensionExecuteTests**.cpp adds tests to verify the Execute API for these column types.
Note : Binary, Varbinary column types are not supported even though such Parameter types are supported. This is in sync with what we have for inbox R.

This PR also has few nit fixes:
1) use default RHOME value in CMakeLists instead of hard coded value
2) removes usage of "using directive" from header files.
3) replace reinterpret_cast with static_cast
4) align function parameters with spaces and match their names with the ExtHost API names.