Merged PR 2757004: Add Python input params tests
Add the input parameter tests for INT, BIT, REAL, FLOAT(n), BIGINT, SMALLINT, TINYINT, CHAR(n), VARCHAR(n), BINARY(n), VARBINARY(n)

As part of this change, we need to add boost python to the test project (originally it was only needed in the actual extension). This is because we need to access the boost python namespace to confirm whether the parameters that are created have the correct values.

Also, while writing the tests, I noticed that the binary params did not quite work as expected. Boost::python does not handle binary/bytes by default, so we need to use the underlying Python C++ api to create a memoryview object, then convert it to bytes before assigning it in the python namespace. It is a good preview of how other, more complex objects such as Timestamp will work.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/2757004