Merged PR 2781850: add params container
Add PythonParam and PythonParamsContainer to encapsulate parameters, and for the functions that create and add the parameters to the namespace. By creating these classes we can also make templates for some functions, and prepare for output parameters.

We store the values of the parameters as boost::python objects, which can then be added to the boost python namespace with no extra conversion. All the conversion is handled when creating the PythonParam object, which will be useful for more complex types (the Raw parameter gives an example).

Also refactor the parameter creation to use a function map, from type to function pointer, to lessen code duplication.