Merged PR 2807543: PythonExtension - Add Execute workflow and DataSets
Add **Execute** workflow and change to boost python numpy.

We change the Dataset manipulation to use numpy here instead of using a dictionary of lists for performance reasons.
When creating a dictionary of lists, each element of the dictionary is a PyObject.
When creating a dictionary of numpy arrays, each array is a PyObject that holds all the information; this significantly reduces the number of PyObjects that are created.

In Python, the execution process is as follows:
1. Add the columns to the DataSet class
2. Add the DataSet to the python namespace as a dictionary of numpy arrays, then convert it to a pandas DataFrame
3. Execute the script
4. Convert the OutputDataSet from a pandas DataFrame to a dictionary in the namespace
5. Retrieve the OutputDataset from the namespace as a python dictionary.

We redirect stdout/err to capture the output from the script, then print them to cout and cerr.

Unfortunately, this comes with a problem: the boost and boost python packages provided by Nuget do NOT have python numpy built in. This is a tracked bug by boost: https://github.com/easybuilders/easybuild-easyconfigs/issues/8810

However, this means we have to use the Boost that bigAnalytics exports, because bigAnalytics does the work and builds boost python with the appropriate libboost_numpy library.
In Linux, we have to install numpy first before building boost so that boost_numpy is built properly.

This PR also adds the Execute tests, which cover all the supported types: simple numbers, CHARs, and BINARYs.