Merged PR 2856532: Use numpy from_data to prevent copies
The original way that we added the C++ objects into the namespace resulted in a lot of copies, where we would copy every element of the array into a python object before adding it to a numpy array and adding it to the namespace.

To avoid these copies, we need to use the from_data method in numpy to create the arrays - this uses the given data pointer and reinterprets it into a numpy array.

Changes in this PR include
- Use from_data for numeric and boolean types
- Refactor to use NA map instead of None objects
- Rename logical to boolean to reflect python's typing
- Move catching the python errors (error_already_set) to the top level pythonextension so that we don't need try-catches inside the rest of the code.
- Change to string and binary to initialize an empty ndarray with the correct size and generic Python object type, then fill it in with Python strings or byte objects.

Variable length types cannot use from_data since they do not have a consistent "stride", which is what numpy uses to distinguish elements in its array. Thus, we need to be a bit more creative with the way we create the PyObjects.

A small change related to the NA handling is that we use NAN, which is never true in a comparison, even when compared to itself. This means that in testing,when we want to confirm if a column has the same values as another one, we need to do some special handling for NANs.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/2856532