Merged PR 2893741: [PythonExtension] Add GetResults and corresponding tests
Add GetResults and the associated tests.

As part of this change, I noticed that extracting SQLCHAR from a boolean in the python namespace does NOT work. Thus, we have to make a separate RetrieveBooleanColumnFromDataFrame to accommodate and extract the boolean properly.

Since the data vector was populated during the Execute step, there are not that many changes to the PythonSession/Extension code, we simply have to retrieve all the information to return to ExtHost in the GetResults call.

The tests follow the template laid out by the GetResultsColumn tests, but instead of checking that each result column has the right names and metadata we actually check that the underlying data is what we expect as well. Since we also need the CheckColumnEquality functions in this file, we move those functions from GetResultColumnTests to the overall cpp.

Finally, there are some whitespace changes.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/2893741