Merged PR 2917579: [PythonExtension] use Nones for all nulls
Use Nones for all null values.

In order to avoid using random set values for NULL, we will use the None object for each type. This has a number of repercussions:

1. The entire numpy ndarray data type becomes "object" instead of int32, int64, etc.
This means we do not know the dtype of the array when we are trying to extract the data from python, so we scan through the array for any non-None values and use the python type (instead of numpy dtype, aka int/float/bool/str). These translate to much larger sizes - SQLBIGINT, SQLDOUBLE.

2. If there are NULLs in the column, we have to create an empty numpy array of "object" type and then populate it with values instead of reinterpreting the data with numpy.from_data. This copies all the data into new objects, so causes a performance regression.

3. We need to rewrite the tests to accommodate the Nones. Since the output type will be changed to BIGINT or DOUBLE we need to add them as special cases.

4. NAN and INF in float types are converted to NULL on the way back to SQL.