Merged PR 3003518: [PythonExtension] Fix linux numpy load segfault
Previously, numpy did not load properly when used with the mssql-satellite user in Linux. This is because the numpy .so file does not directly depend on the python .so in its dependency list, despite actually depending on it. Thus, when we tried to initialize numpy, the load would fail because it could not properly load the library.

Without proper initialization, any call to any boost numpy functions would result in a segfault.

I found this stackoverflow answer that had the solution:
https://stackoverflow.com/questions/49784583/numpy-import-fails-on-multiarray-extension-library-when-called-from-embedded-pyt

What we do now is dlopen the python manually before any initialization, which tells numpy where to find the python library.

There is also a modification to the restore-package.sh file so we install numpy and pandas to the right location, so that the mssql-satellite user has access. Without this change, the packages were installed into /usr/local, which is not the accessible as mssql-satellite.