Merged PR 3149618: Add NVarChar support to Output Params
- Add nvarchar support to output params
- Add wstring test into output params

For output parameters, since the user has to specify the type of the parameter when they create the spees call (nvarchar) we know it is a wstring. Thus we can extract the string from Python as UTF16 and return it directly.

In our wstring test, we output some variation of characters, including a chinese string to test Unicode. To make sure the output parameter is what we expect, we compare the output wstring to the expected wstring byte by byte and make sure everything is the same.