Merged PR 3192747: Add tests for wstring output result columns
Add tests in Execute, GetResults, and GetResultColumn for wstring.

Because Python is a UTF-8 language, we expect the output result columns to always be a normal String as opposed to a WString. This means that, in our tests, we have to input wstrings but check for normal strings in the output. Because the code already assumes that all python "str" objects are strings (not wstrings), there are no changes to the extension code.

For Unicode characters, this means that some garbage characters will be returned, while normal characters will be returned properly.

Because we use the -fshort-wchar flag when compiling Linux, many of the built-in functions for wchar/wstrings do not work (such as wcslen). This is why we hardcode the sizes in the strLen parameters.