Merged PR 3246751: [PythonExtension] Datetime in Param
Adding Date and Datetime support in the Python Extension for Params / Output Params.

In order to use the Python API for Date/DateTime, we need to include the datetime.h header which is not included in the normal Python.h header. After including datetime.h, we have to use PyDateTime_Import which is a macro that #DEFINEs all the PyDate functions. Without this Import we would segfault.

Also added tests to test the same.

We check Pandas Timestamp objects as well as datetime.datetime objects to make sure both work, since the user could prefer one or the other.

We do not actually support TIME in ExtHost, so we don't have any tests for it. I am not sure why this is the case, but when I try running anything in the extension with Time SQL type it returns an error:

Parameter '@TimeTest' uses a data type that is not supported by the runtime for 'PyExtension' script. Unsupported types are timestamp, datetime2, datetimeoffset, time, text, ntext, image, hierarchyid, xml, sql_variant and user-defined type.