Merged PR 3256384: [PythonExtension] Add Date support to DataSet and Columns
Add DateTime and Date support in DataSet columns.

DateTime is represented by SQL_C_TYPE_TIMESTAMP and SQL_TIMESTAMP_STRUCT.

Also added tests for the same.

When we pull out the dtype (dataType) from the numpy column in the pandas dataframe, timestamp objects are typed as "datetime64[ns]". You'll notice that in the type map we also set that as a SQL_TIMESTAMP_STRUCT, because they are equivalent. Numpy automatically makes this conversion inside python.

To avoid this type (which is held as a floating point number of nanoseconds from 1970/1/1), we extract the column from pandas as a list instead of a numpy array like we do for all the other types (line 1147 PythonOutputDataSet::RetrieveDateTimeColumnFromDataFrame). This allows us to use the PyDateTime API to extract all the timestamp information.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3256384