Merged PR 3281829: [PythonExtension] Add C++17 compiler, use if constexpr where possible
Upgrade the compiler from c++14 to c++17 in Python.

This is useful for the "if constrexpr" construction, which is used in template functions to distinguish between different specializations of the template.

As part of this, we download a newer version of gtest  that supports c++17, and install gcc-7 in linux.

c++17 also allows us to replace is_same<>::value  to is_same_v which does the same thing.

Note we keep the older version of gtest as well, because R requires c++14 and that associated version of c++.