Merged PR 3403887: [RExtension] Update to use R-4.0.2, Rtools 4.0 with updated GNU compiler and std c++17
C++17 provides use of new constructs like **constexpr** useful for better templatized coding.
So its beneficial to upgrade. That means g++ compiler needs upgrade.

On linux, just specifying C++17 is sufficient.
But on windows, RExtension uses the gnu compiler from Rtools. This means Rtools needs an update to use new compiler. To use new Rtools, we need to update R itself to R-4.0.2

Hence, this change is to update RExtension, RExtension-test to build using R 4.0.2, Rtools 4.0 on windows, c++17. Similarly, the googletest that is built in the repo is updated to c++17. Previous attempt to update R to 4.0.2 was failing to link RInside because Rtools 4.0 was not being used back then.
We still use Rtools 3.5 for mingw32-make since that is no longer provided in Rtools4.0

With this update, the use of 'print' function in the test produces junk character in the beginning and end of the printed strings on Windows.
e.g. with print, [1] "Hello RExtension!" is getting printed as : [1] "ÿ_Hello RExtension!ÿ_".
This happens only during unit test and not when tested end-to-end via sqlserver. So 'print' is replaced by 'message' which doesn't have this problem.
However, on Linux, output from "message" isn't captured by testing::internal::CaptureStdout(); causing the test to fail. So we preserve use of 'print' on Linux for the string that we check in the test. Other output strings that we don't test have been moved to use "message" even on Linux.

[Task 864391: Upgrade to C++17, R 4.0.2](https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/864391)

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3403887