Merged PR 3452565: [RExtension] Add NCHAR/NVARCHAR support for Output parameters
This PR is a duplicate of the already approved (delta) changes for Output NVARCHAR parameters but now going into master.
https://msazure.visualstudio.com/One/_git/Data-SQL-Language-Extensions/pullrequest/3433449

This PR adds support for nchar/nvarchar output parameters by modifying **RTypeUtils::FillDataFromCharacterVector** - templatizing it to fill the character data vector of type SQLType - SQLWCHAR for nchar/nvarchar and SQLCHAR for char/varchar. R stores both types of character data in Rcpp::CharacterVector in utf-8 encoding. We convert utf-8 into utf-16 when it is SQLWCHAR before sending it back to Exthost via the returned paramValue.
The type of RCharacterParam::m_value is also modified to be SQLWCHAR/SQLCHAR.
The number of bytes occupied by the parameter value in utf-16 encoded format are returned in strLenOrInd hence we multiply the length by sizeof(SQLType).

We also add tests for the same to **RExtensionGetOutputParamTests.cpp**

[Task 792769: Add NVarChar / NChar support](https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/792769)