Merged PR 3459485: [RExtension] Add Date/Datetime support for Input parameters in InitParam
This PR adds support for **date**,**datetime**, **datetime(2)** data types in **InitParam**.
We create a templatized **RDateTimeParam** class to handle these types and create the required RcppVector in new function **CreateDateTimeVector** in RTypeUtils.

Date is represented as SQL_DATE_STRUCT as the SQLType and RType as Rcpp::DateVector containing the Rcpp::Date as the type for individual elements.
Similarly, datetime is represented as SQL_TIMESTAMP_STRUCT as the SQLType and RType as Rcpp::DatetimeVector containing Rcpp::Datetime as the type for individual elements.
Handling of nanoseconds is done by representing them in seconds after decimal point.

Support for output parameters and datasets will be in a subsequent PR.

We also add tests for input date and datetime parameters in **RExtensionInitParamTests**.

And to remove redefinition warnings ERROR and ODBCVER are undefined in the header.

[Task 792770: Add DateTime Support](https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/792770)

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3459485