Merged PR 3472985: [RExtension] Add Date(time) support for Input Dataset
This PR adds support for date,datetime, datetime2 data types to InputDataSet.
We add the templatized function "**AddDateTimeColumnToDataFrame**" to handle these types and create the required RcppVectors for each column using the function **CreateDateTimeVector** in RTypeUtils.

Date is represented as SQL_DATE_STRUCT as the SQLType and RType as Rcpp::DateVector containing the Rcpp::Date as the type for individual elements.
Similarly, datetime is represented as SQL_TIMESTAMP_STRUCT as the SQLType and RType as Rcpp::DatetimeVector containing Rcpp::Datetime as the type for individual elements.
Handling of nanoseconds is done by representing them in seconds after decimal point.

Support for output dataset will be in a subsequent PR.

We also add execute tests for InputDataSet with date(time) columns and executing script that assigns InputDataSet to OutputDataSet. We verify that the DateVector columns contain the expected data as well as the fact that the input and output data frames are equal.

Additionally, the static member m_rowsNumber is renamed to sm_rowsNumber in ColumnInfo class.
Also move back to using print to see the correct DateTime values in the stdout (albeit some random characters before and after the value on Windows since the GUI is not set initialized)

[Task 792770: Add DateTime Support](https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/792769)

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3472985