Merged PR 3529554: [RExtension] Fix error log showing cant find TZDIR environment variable length
**Issue**:
We used the same function to **GetAndLogLastError**. We want to log error when we are getting R_HOME on Windows and if we can't find it.
But we do not want to show error while getting environment variable if we are going to set it as well such as in case of R_HOME on linux and TZDIR.
**Fix**:
To achieve this, modify **GetEnvVariable** to accept the logError argument and split the function GetAndLogLastError into two and call them only when we want to log errors.

Bug:
https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/900379

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3529554