Merged PR 3512820: [PythonExtension] fixes with DateTime
Fix a problem with DateTime.

Numpy datetime has a special NULL value called NaT (not a time). This was not handled by boost's isnan function, and we need to check it ourselves.

In the DataFrame, the type is NaTType so we can easily check that and change those to NULLs. In params, we need to use numpy.isnat to check whether the number is NaT because the type is considered to be np.datetime64.

Added some tests to test the same.

Also fixed one major bug in the tests where we were not actually comparing the expected timestamp values and the actual timestamp values.