Merged PR 3542833: [RExtension] Fix bit values for 0 and 1
Issue:
Bit value anything other than '0' was being considered as a TRUE logical in R. Note '0' and 0 are different.

Fix:
Add a condition to check for 0 in addition to '0' to denote FALSE in R. All other values are considered TRUE.
While retrieving data from R, FALSE should be set as 0 instead of '0'; and TRUE as 1 instead of '1' since SQL interprets both '0' and '1' as 1 BITs.

Tests:
Add bit value tests for 0 and 1 for parameters and datasets.

Bug:
https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/900410

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3542833