Merged PR 3545534: [RExtension] Honor nullable column information
**Issue**:
We were focussed on checking for strLenOrInd value = SQL_NULL_DATA for the input column of a dataset to be NULL.  However, if nullable is SQL_NO_NULLS, we should disregard whatever the value of strLenOrInd is.

**Fix**:
Check if the column's nullable value is SQL_NULLABLE in addition to strLenOrInd!=nullptr and
strLenOrInd[0] == SQL_NULL_DATA to conclude the value is NULL aka NA in R.
If nullable is SQL_NO_NULLS, we disregard strLenOrInd value.

**Test**:
Added a new parameter to InitializeColumn to indicate nullable.
In all the tests, with same data type, column1 is SQL_NO_NULLS and column2 is SQL_NULLABLE. Use ColumnInfo to store this information.

In RExtensionExecuteTests, for different data type columns, the DoubleColumn of the input dataset is NOT nullable i.e. SQL_NO_NULLS. For testing that strLenInd will be ignored, intentionally pass in a non-null strLenOrInd with some SQL_NULL_DATA when calling Execute. While verifiying, make sure we didn't set any row of the DoubleCol to NA.

Bug:
https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/904321

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3545534