Merged PR 3560036: [RExtension] Add more library management tests, fix for correct install location check
This PR adds more tests to the **RExtensionLibraryApiTests** suite.

It also fixes the check to find if a library is installed by trying to find it in the desired location and concluding if it is installed or not if the found location matches the desired location.
Note the package may be installed somewhere else in the libPath in which case found location will not be same as desired location. The comparison done is using string compare after normalizing the path since fs::equivalent was unreliable as proven by tests.

We also add different packages to test different scenarios.

https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/906895

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3560036