Merged PR 3596577: [RExtension] Return sizeof(SQLCHAR) as the min columnSize for output character column
**Issue**:
For a character column, Exthost needs the result character columnSize to be at least sizeof(SQLCHAR) -> even when the column contains all NAs or empty strings.
In R, having NAs in a column, is by default, treated as a logical column. So this issue is specific only for character columns.

If the OutputDataSet has a character column with all null values in it or all empty strings in it, and if we don't send the columnSize to be sizeof(SQLCHAR), instead send it as 0, Exthost errors out - aborting the SPEES since Exthost/SQLServer  interprets this columnSize to be the length of the character data type - the n in char(n). So sending columnSize = 0 is interpreted as char(0) which is not allowed and so Exthost errors.

This is not an issue with raw or binary column either since NA is treated as 0x00 raw bytes in R. So the following succeeds :
`EXEC sp_execute_external_script
	@language =N'myR',
	@script=N'
OutputDataSet <- data.frame(rawNA = as.raw(c(NA,NA)));'
WITH RESULT SETS ((rawCol binary));`

**Fix**:
Initialize the columnSize to be sent back for character columns with sizeof(SQLCHAR).

**Test**:
Did end to end testing as well as added new unit tests.
`EXEC sp_execute_external_script
    @language =N'myR',
    @script=N'
OutputDataSet <- data.frame(logicalNA = as.logical(c(NA,NA)), charNA = as.character(c(NA, NA)), emptyChar = as.character(c("", "")));'
WITH RESULT SETS ((logicalCol BIT, nullCharCol VARCHAR(5), emptyCharCol VARCHAR(5) NOT NULL));`

https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/916743

Also fixed a warning and new line/file name issues backporting from oss.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3596577