Merged PR 3601864: [PythonExtension] Fix empty string nullptr output param
**Issues**:

SPEES using PythonExtension with an empty char/binary output parameter would error out in Exthost with the error E_INVALID_PROTOCOL_FORMAT in VerifyOutputParam.

**Root cause:**
We were returning strLenOrInd != SQL_NULL_DATA && paramValue == nullptr. Value of strLenOrInd being passed back = 0 is correct. However, paramValue should not have been nullptr, instead should contain 1 element which is an empty zero-length string.

**Fix**:
We insert the \0 value into the data vector when the length of the output parameter is 0. This causes the data vector size to be 1, but we save the strLen or Ind as 0 so that ExtHost gets the correct information about the underlying string.