Merged PR 3617575: [RExtension] Fix Output Char Param for Empty strings
**Issues**:

1. SPEES using RExtension with output character parameter having an empty string as its value after the R script execution would error out in Exthost with the error E_INVALID_PROTOCOL_FORMAT in VerifyOutputParam.
2. Unit tests failed to catch this issue even though the scenario was added to the test.

**Root cause:**
For 1, we were returning strLenOrInd != SQL_NULL_DATA && paramValue == nullptr. Value of strLenOrInd being passed back = 0 is correct. However, paramValue should not have been nullptr, instead should contain 1 element which is an empty zero-length string.
This was due to the following way of inserting a 0 length string into a vector:
data->insert(data.end(), emptyString.begin(), emptyString.begin() + lengthOfString(=0));

For 2, the root cause was that we were not checking if the returned paramValue was nullptr or not. We were instead constructing a string from the returned paramValue which would become an empty string since we were returning nullptr and this would match the expectedParamValue String. And this masked the paramValue being passed back as nullptr for empty strings.

**Fix and test**:
1. Insert the target string in the data vector based on the number of rows and the length of the string.
For multiple rows with interspersed empty strings in the data, don't add 0 length strings
between two non-empty strings. This is the case for a dataset having char column with multiple rows and at least one of them is an empty string.
For a single row with empty string, add a 0 length string explicitly. This is the case of output char parameter or a dataset char column with single row. This way we avoid the vector being of size 0 and don't return a nullptr as the paramValue.

2. Fix the unit tests by checking whether the returned paramValue is nullptr or not before constructing the string for comparison with expectedParamValue strings.

**Tested end-to-end :**
		 `DECLARE @param1 CHAR(2)`
		 `DECLARE @param2 VARbinary(2)`
  `EXEC sp_execute_external_script
   @language =N'myR',
   @script=N'
   param1 <- c("");
  param2 <- as.raw(NA);
  print(param1);
   print(param2);'
,@params = N'@param1 CHAR(2) OUTPUT, @param2 VARbinary(2) OUTPUT'
,@param1 = @param1 OUTPUT
,@param2 = @param2 OUTPUT`
`SELECT @param1 as param1, @param2 as param2`

**Note**:
The issue doesn't exist with (var)binary output parameter/dataset column

https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/915231

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3617575