Merged PR 3788002: [RExtension] Add numeric column InputDataSet support for RExtension
This change is to add a new function **AddNumericColumnToDataFrame** to support numeric(p,s)/decimal(p,s) types in InputDataSet.
A unit test is also added to ExecuteApiTests where we confirm the InputDataSet can be read as expected in the R script.

Note: There is no need to add code/test for numeric columns in OutputDataSet in the extension since even if the class in R is numeric, it is mapped to double i.e. SQL_C_DOUBLE type by default.
If the WITH RESULT SETS clause in the SPEES specifies a numeric(p,s) column output type - SQL will do the conversion from double to numeric and return correct results if the double is in range of the given numeric, or error out if its not in range.
This limitation is because the extension doesn't know the output schema in WITH RESULT SETS and hence by default maps the numeric class in R to double which is a superset of SQL's numeric(p,s) type.

[Task 959365: Add Numeric Column in InputDataSet Support](https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/959365)

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3788002