Merged PR 3788226: [PythonExtension] Add changes for NVarchar handling (encoding) and DateTime
Similar to this change for parameters (https://msazure.visualstudio.com/One/_git/Data-SQL-Language-Extensions/pullrequest/3747968) we need to make a change in the way we handle nvarchar in DataSets, because of encoding.

The original function "PyUnicode_FromKindAndData" did not work with encoded UTF-8 strings properly. For char values, it would only work if all chars were 1 byte.

The new functions PyUnicode_DecodeUTF8/16 specifically take UTF8 or 16 encoded strings and create PyUnicode objects.

I also noticed that DateTime was not returning the correct nanosecond values, because I forgot to set the DecimalDigits parameter for Timestamp objects.