Merged PR 3855747: [PythonExtension] Change return types for NULL / object type columns
Modify return values for "object" and other columns with NULL values. In particular, binary columns are always considered 'object' columns in python, so they are always returned as SQL_C_CHAR. We initialize datetime columns as object columns so that they are not bound by the pandas timestamp restrictions.

These changes are necessary for an "all-null" column (we return strings) and for initializing columns with the correct column type. NULLs in number columns are automatically modified into NaN (which is a float64 in numpy).

We also needed to modify a number of the unit tests to match these new changes.

DS_MAIN_DEV change here: https://msdata.visualstudio.com/Database%20Systems/_git/DsMainDev/pullrequest/470396
Note that the DS_Main_Dev PR can't really be checked in or tested right now because we need to update the release of the Data-Sql-Language package on it, so it is still a Draft PR. I have tested it manually and it works.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3855747