Merged PR 3866585: [RExtension] Fixing DateTime issues discovered by TestShellTests
Developing OutputParamIsm and InputDataQueryIsm TestShellTests for RExtension, two DateTime issues came up:

1) Providing DateTime/DateTime2 data to SPEES with a mix of NULL and non-NULL values, we've been getting _Msg 9803, Level 16, State 3, Line 53, Invalid data for type "datetime2"_.
Debugging the issue, the data that were returned to SqlServer after a NULL-valued was garbage and that's why SqlServer couldn't convert it to DateTime2. In the case of NULL for DateTime(2) data types, we need to add a zero timestamp to the data vector, this is needed since the ODBC protocol used for exchange of data between Exthost and the extension expects empty value for null if the type is fixed and, timestamp is a fixed type.

2) With hard-coded decimal digits 0 in RDataSet.cpp, we've been always returning zero as fractional seconds for the DateTime variant. In SQL Server, the default is 7 (https://docs.microsoft.com/en-us/sql/t-sql/data-types/datetime2-transact-sql). Execution of TestShellTests showed us beyond 6 digits, R runtime, and SQL Server mismatch (some discussions here: https://stackoverflow.com/questions/7726034/how-r-formats-posixct-with-fractional-seconds).

The unit tests were updated to reflect these changes. In particular,
1) Since in the case of NULL for DateTime(2), we return value 0 (non-null), we've updated  RExtensionGetOutputParamTests for DateTime to check equality to zero instead of expecting a nullptr.

2) Updating decimal digits 0 to 6, RExtensionGetResultColumnTests for DateTime (i.e. GetDateTimeResultColumnsTest) were updated to expect 6 decimal digits.

**Test**
Unit tests were passed in local.

PS: this is a pre-requisite PR for a PR in DSMainDev which is not published yet.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3866585