Merged PR 3878569: [RExtension]: Interpret NULL binary/varbinary as row(0)
Previously, we've returned 0x00 in case of a NULL (var)binary since a raw vector of length 1 with value = NA coerced to be 0 (the nul byte).  In terms of getting length 1, this was consistent with other data types. However, this means when we got back the input parameter as an output of the SPEES we've been getting it back as 0x00 – causing the NULL to change to 0x00.

This changeset handles this case by not pushing-back anything into the RawVector when byte value is NULL.
In creating a raw vector, if strLen_or_Ind is nullptr, returns an empty RawVector. Otherwise, iterates over data and for each element, if strLen_or_Ind at any index is not SQL_NULL_DATA, go over each of its byte to push_back the value to the vector. If strLen_or_Ind at any index is SQL_NULL_DATA, does not push_back anything.

Sending back an empty raw vector with zero row numbers and zero column size makes a case:
If there is no row in the DataFrame, columnSize = 0. In case of XVT_VARBYTES, ExthostExtensionManager expects to get a columnSize > 0, otherwise, GetOdbcCTypeInfo considers it as an invalid column and raises E_INVALID_PROTOCOL_FORMAT. We thus updated columnSize = 1 in this case.

The tests were updated accordingly. For the expectedParamValue NULL cases:
1) In RExtensionInitParamTests.cpp, now, we expect to get param.size() == 0 instead of param itself to be zero.
2) In RExtensionGetResultColumnTests.cpp, in GetRawResultColumnTest, the expected columnSize is 1;