Merged PR 3901578: [PythonExtension] Remove vector construction from string and raw
Remove vector construction from string and raw because of vector memory leak.

Modify the way we retrieve string columns.
We used to create a vector<char> * and push all the string data onto it, then return the underlying data pointer. Unfortunately, this causes issues when we try to free the data. Stealing the data pointer out from under the vector and then deleting it ourselves is not supported, and seems to cause an exception when deleting. However, since delete itself will never throw, we don't hit this issue in the non-streaming case (the process is destroyed immediately after so we don't have to deal with the corrupt memory). In the streaming case though, we keep running the process and there is an exception when we try to work with the memory in the second chunk. Furthermore, the vector object itself (not the data underneath) is also allocated on the heap, and we never delete it, so we are leaking memory.

https://msdata.visualstudio.com/Database%20Systems/_sprints/taskboard/SQL%20Extensibility/Database%20Systems/Azure%20Semesters/Iron%20-%20CY2020%202H/2020-12?workitem=991529