Merged PR 3907824: [RExtension] Remove vector construction and resolve its memory leakage
Remove vector construction from all data types because of the vector memory leak.
We have been stealing the data pointer out from under the vector. The vector object itself (not the data underneath) is also allocated on the heap, and we never delete it, so we are leaking memory.

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Data-SQL-Language-Extensions/pullrequest/3907824