Merged PR 3925500: Don't try signing .so files in the Linux pipeline
**Linux build suddenly was failing after 12/15/2020 on signing issues**
On 12/15/2020 `OneBranch Pipeline` team shipped support for ESRP signing on Linux.

Our had signing configured in our CDPx Linux pipeline YAML prior to 12/15/2020. This was actually a no-op and we were not actually signing any files. Prior to 12/15/2020, we could only sign with the addition of a Windows phase and a CDPx Windows pipeline YAML.

When Linux Signing was shipped to CDPx production **on 12/15/2020**, the Linux signing configuration in the Linux pipeline YAML **became active** and **began signing files** in the Linux phase. _Our signing was incorrectly configured (e.g. signing files that cannot be signed like .so file in here)_ then **this started to fail the build**.

To remediate this, we **turned off signing on Linux completely by removing all the signing_options properties in the Linux YAML**. Since signing was a no-op prior to 12/15/2020, removing all the signing_options properties will effectively put us back into the pre-12/15/2020 state.

More info:
https://onebranch.visualstudio.com/Pipeline/_wiki/wikis/Pipeline.wiki/4679/Linux-Signing?anchor=why-is-my-linux-build-suddenly-failing-after-12%2F15%2F2020-on-signing-issues%3F